changed TileChangeRequest handling added lock access, unaccess fixed re-apply lock empty area fixed role reseting fixed no name shown when punching locks and mailbox added bulletin board fixed drop causing crash fixed Telnet auth bug fixed register dialog not shown

+++ Source/World/TileInfo.h
@@ -75,7 +75,7 @@ public:
     bool HasFlag(uint16 flag) { return m_tileData->flags & flag; };
     void ToggleFlag(uint16 flag) { m_tileData->flags ^= flag; }
 
-    void PunchTile(uint8 damage);
+    void PunchTile(uint32 damage);
     bool WillBreak(uint8 damage);
     float GetHealthPercent();
